% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_bandwidths_get_bw.R
\name{add_bw}
\alias{add_bw}
\title{Add a new bw to \code{bw_environment}.}
\usage{
add_bw(bw_str, bw)
}
\arguments{
\item{bw_str}{A string giving the name of the density.}

\item{bw}{The bw function.}
}
\value{
None.
}
\description{
Add a new bw to \code{bw_environment}.
}
\keyword{internal}
