% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{add_kernel}
\alias{add_kernel}
\title{Add a new kernel to \code{kernels_environment}.}
\usage{
add_kernel(kernel_str, kernel)
}
\arguments{
\item{kernel_str}{A string giving the name of the density.}

\item{kernel}{The kernel function.}
}
\value{
None.
}
\description{
Add a new kernel to \code{kernels_environment}.
}
\keyword{internal}
