% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth_functions.R
\name{bandwidth_selectors}
\alias{bandwidth_selectors}
\alias{bw.JH}
\alias{bw.RHE}
\title{Bandwidth Selectors for Kernel Density Estimation}
\usage{
bw.JH(x, kernel = NULL, start = NULL, support = NULL)

bw.RHE(x, kernel = NULL, start = NULL, support = NULL)
}
\arguments{
\item{x}{numeric vector containing the data.}

\item{kernel}{(optional) a \code{kernel} in list format.}

\item{start}{(optional) a \code{start} in list format.}

\item{support}{(optional) the support of the data.}
}
\description{
Bandwidth selectors for \code{kdensity}.
}
\details{
Bandwidth functions for parametric starts and asymmetric kernels currently include:

\itemize{
  \item \strong{bw.JH:} Selector for the Gaussian copula kernel, based on
  normal reference rule. Described in Jones & Henderson. The default method when
  the \code{gcopula} kernel is used in \code{kdensity}.
  \item \strong{bw.RHE:} Selector for parametric starts with a symmetric kernel.
  Described in Hjort & Glad. The default method in \code{kdensity} when a parametric
  start is supplied and the kernel is symmetric.
}

In addition to these,\code{kdensity} supports the bandwidth functions described in
\code{\link[stats]{bandwidth}}.
}
\seealso{
\code{\link{kdensity}}, \code{\link[stats]{bandwidth}} for the
bandwidth selectors for \code{\link[stats]{density}}.
}
