% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{get_kernel_sd}
\alias{get_kernel_sd}
\title{Helper function that gets a the kernel standard deviation.}
\usage{
get_kernel_sd(kernel)
}
\arguments{
\item{kernel}{a string specifying which kernel to use.}
}
\value{
a kernel function of the format k(u) with standard deviation
normalized to 1.
}
\description{
Helper function that gets a the kernel standard deviation.
}
\details{
stats::density use this concept in order to make different
kernels directly comparable. Most of the bandwidth selectors are based
one the Gaussian kernel, and translation requires the kernels to be
comparable.
}
