% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_handlers.R
\name{get_support}
\alias{get_support}
\title{Helper function that generates automatic support.}
\usage{
get_support(start_str, kernel_str)
}
\arguments{
\item{start}{the type of parametric start.}

\item{kernel}{the type of kernel.}
}
\value{
a support proposal. Currently kernel does nothing.
}
\description{
Helper function that generates automatic support.
}
