% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_functions.R
\name{kdensity_start_functions}
\alias{kdensity_start_functions}
\title{Get densities and estimators from strings.}
\usage{
kdensity_start_functions(start, support)
}
\arguments{
\item{start}{a string specifying the density of interest.}

\item{support}{a tuple containg the support. Used for error handling.}
}
\value{
a list of two functions.
}
\description{
Get densities and estimators from strings.
}
