% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{listmerge}
\alias{listmerge}
\title{Puts default arguments into ellipses: ...}
\usage{
listmerge(x, y, type = c("merge", "template"))
}
\arguments{
\item{x}{a list of default arguments.}

\item{y}{a list of supplied arguments}

\item{type}{should x and y be merged (with y having priority),
or the elements x be a template filled with values from y?}
}
\value{
a merged list where conflicts are solved in favour
of supplied.
}
\description{
Puts default arguments into ellipses: ...
}
