% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_starts_custom_maximum_likelihood.R
\name{mlbeta}
\alias{mlbeta}
\title{Fast maximum likelihood for the beta distribution.}
\usage{
mlbeta(x, start = NULL, type = c("none", "gradient", "full"))
}
\arguments{
\item{x}{The data, contained in the unit interval.}

\item{start}{Optional start point for the \code{nlm} function.}

\item{type}{Should we use a precomputed gradient function or Hessian?}
}
\value{
a vector of two parameters, shape1 and shape2.
}
\description{
Fast maximum likelihood for the beta distribution.
}
\details{
The option "none" is fastest at the moment. The algorithm is
a straight forward implementation of maximum likelihood for beta using
the sufficient statistics.
}
