% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_maximum_likelihood.R
\name{mlgamma}
\alias{mlgamma}
\title{Fast maximum likelihood for the gamma distribution.}
\usage{
mlgamma(x, rel_eps = 10^-10, max_iter = 100)
}
\arguments{
\item{x}{The data, contained in c(0, Inf)}

\item{rel_eps}{Relative epsilon comparison criterion.}

\item{max_iter}{Maximal number of Newton-Raphson iteratons.}
}
\value{
a vector of two parameters, \code{shape} and \code{rate}.
}
\description{
Fast maximum likelihood for the gamma distribution.
}
\details{
Implemented following wikipedia.
}
