% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.kdensity}
\alias{plot.kdensity}
\title{Plotting generic for kdensity objects.}
\usage{
\method{plot}{kdensity}(obj, range = NULL, ...)
}
\arguments{
\item{obj}{a kdensity object}

\item{range}{the x-values the density is applied over (optional)}

\item{...}{passed to plot.default.}
}
\description{
Plotting generic for kdensity objects.
}
