% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{recycle}
\alias{recycle}
\title{Recycles arguments.}
\usage{
recycle(..., prototype)
}
\arguments{
\item{...}{A list of arguments to be recycled.}

\item{prototype}{an optional argument. If given, repeats all arguments
up to the length of the prototype. If an element of the list has the name,
it is used. If not, the variable itself is used.}
}
\description{
Recycles arguments.
}
\details{
Recycles arguments so that all vectors are equally long. If a
prototype is given, each vector will have the same size as the prototype.
}
\keyword{internal}
