% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{support_compatible}
\alias{support_compatible}
\title{Checks compatibility between supports.}
\usage{
support_compatible(kernel, start, support)
}
\arguments{
\item{kernel, start, support}{The kernel, start and support to check.}
}
\value{
None.
}
\description{
The supplied support must never be larger than the support of
the parametric start / kernel.
}
\keyword{internal}
