% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_starts_get_start.R
\name{add_start}
\alias{add_start}
\title{Add a new parametric start to \code{starts_environment}.}
\usage{
add_start(start_str, start)
}
\arguments{
\item{start_str}{A string giving the name of the density.}

\item{start}{The parametric start function.}
}
\value{
None.
}
\description{
Add a new parametric start to \code{starts_environment}.
}
\keyword{internal}
