% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_bandwidths.R
\name{bandwidths}
\alias{bandwidths}
\alias{bw.JH}
\alias{bw.RHE}
\title{Bandwidth Selectors}
\description{
Bandwidth selectors for \code{kdensity}. These are the available options
for the option \code{bw} in \code{kdensity}.
}
\section{Bandwidth selectors}{

   \code{"nrd0", "nrd", "bcv", "SJ"}: Bandwidth selectors from \code{stats}.
   They are documented in \code{\link[stats]{bandwidth} stats:bandwidth}.
   "nrd0" is the standard bandwidth selector for symmetric kernels with
   constant parametric starts.


  \code{"ucv"}: Unbiased cross validation. The usual standard option for
  asymmetric kernels.


  \code{"JH"}: Selector for the Gaussian copula kernel, based on
  normal reference rule. Described in Jones & Henderson. The default method when
  the \code{gcopula} kernel is used in \code{kdensity}.


  \code{"RHE"}: Selector for parametric starts with a symmetric kernel,
  based on a reference rule with Hermite polynomials.
  Described in Hjort & Glad (1995). The default method in \code{kdensity} when a parametric
  start is supplied and the kernel is symmetric.
}

\section{Structure}{

   The bandwidth selector is a function of four arguments: The data
   \code{x}, a kernel string \code{kernel}, a start string \code{start},
   and a support vector \code{support}. To obtain the functions associated
   with these strings, use \code{get_kernel} and \code{get_start}. The
   function should return a double.
}

\examples{
   ## Not a serious bandwidth function.
   silly_width = function(x, kernel = NULL, start = NULL, support = NULL) {
     kernel = get_kernel(kernel)
     kernel$kernel(0.5)
   }
}
\references{
Jones, M. C., and D. A. Henderson. "Kernel-type density estimation on the unit interval." Biometrika 94.4 (2007): 977-984.

Hjort, Nils Lid, and Ingrid K. Glad. "Nonparametric density estimation with a parametric start." The Annals of Statistics (1995): 882-904.
}
\seealso{
\code{\link{kdensity}}, \code{\link[stats]{bandwidth}} for the
   bandwidth selectors of \code{\link[stats]{density}}. In addition,
   \code{\link{kernels}}; \code{\link{starts}}
}
