% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_bandwidths_get_bw.R
\name{get_bw}
\alias{get_bw}
\title{Get bandwidth functions from string.}
\usage{
get_bw(bw_str)
}
\arguments{
\item{bw_str}{a string specifying the density of interest.}
}
\value{
a bandwidth function.
}
\description{
Get bandwidth functions from string.
}
\keyword{internal}
