% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_kernels_get_kernel.R
\name{get_kernel}
\alias{get_kernel}
\title{Helper function that gets a kernel function for kdensity.}
\usage{
get_kernel(str)
}
\arguments{
\item{str}{a string specifying which kernel to use.}
}
\value{
a kernel function of the format k(u) with integral normalized
to 1.
}
\description{
Helper function that gets a kernel function for kdensity.
}
