% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdensity_helpers.R
\name{get_kernel_start_support}
\alias{get_kernel_start_support}
\title{Fill in missing kernel, start or support given the supplied values.}
\usage{
get_kernel_start_support(kernel, start, support)
}
\arguments{
\item{kernel}{Supplied kernel; string or list.}

\item{start}{Supplied parametric start; string or list.}

\item{support}{Binary vector.}
}
\value{
a list with members kernel, kernel_str, start, start_str,
and support.
}
\description{
This function takes the supplied values of kernel, start, and support
and fills in the non-supplied ones. It also handles inconsistencies,
such as providing a support on (-Inf, Inf) but a kernel on (0, Inf).
}
\details{
The \code{kernel} and \code{start} parameters are either strings or
adhering to the kernel/start list structure. \code{support} is a
numeric vector of length two.
}
\keyword{internal}
