% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_range}
\alias{get_range}
\title{Supplies a plotting range from a kdensity object.}
\usage{
get_range(obj)
}
\arguments{
\item{obj}{A kdensity object.}
}
\value{
S vector of size 1000, used for plotting.
}
\description{
Supplies a plotting range from a kdensity object.
}
\keyword{internal}
