% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth_functions.R
\name{get_standard_bw}
\alias{get_standard_bw}
\title{Get a bandwidth string when 'bw' is unspecified.}
\usage{
get_standard_bw(kernel_str, start_str, support)
}
\arguments{
\item{kernel_str}{supplied kernel,}

\item{start_str}{supplied parametric start}

\item{support}{supplied support}
}
\value{
a bandwidth string.
}
\description{
Get a bandwidth string when 'bw' is unspecified.
}
