% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_bandwidths_get_bw.R
\name{get_standard_bw}
\alias{get_standard_bw}
\title{Get a bandwidth string when 'bw' is unspecified.}
\usage{
get_standard_bw(kernel_str, start_str, support)
}
\arguments{
\item{kernel_str}{a kernel string}

\item{start_str}{a parametric start string.}

\item{support}{the support.}
}
\value{
a bandwidth string.
}
\description{
Get a bandwidth string when 'bw' is unspecified.
}
\keyword{internal}
