% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_functions.R
\name{get_start}
\alias{get_start}
\title{Get densities and estimators from strings.}
\usage{
get_start(start, support)
}
\arguments{
\item{start}{a string specifying the density of interest.}

\item{support}{a tuple containg the support. Used for error handling.}
}
\value{
a list of two functions.
}
\description{
Get densities and estimators from strings.
}
