% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_starts_get_start.R
\name{get_start}
\alias{get_start}
\title{Get densities and estimators from strings.}
\usage{
get_start(start_str)
}
\arguments{
\item{start_str}{A string specifying the density of interest.}
}
\value{
A list of two functions.
}
\description{
Get densities and estimators from strings.
}
\keyword{internal}
