% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdensity.R
\name{kdensity}
\alias{kdensity}
\title{Kernel density estimator with parametric start}
\usage{
kdensity(x, adjust = 1, support = NULL, na.rm = TRUE, normalized = TRUE,
  bw = c("nrd0", "nrd", "bcv", "ucv", "SJ"), kernel = c("gaussian",
  "epanechnikov", "rectangular", "triangular", "biweight", "cosine",
  "optcosine", "tricube", "triweight", "laplace"), start = c("uniform",
  "normal", "gamma", "exponential", "inverse_gaussian", "lognormal", "beta",
  "laplace"))
}
\arguments{
\item{x}{the supplied data.}

\item{adjust}{an adjustment constant, so that h = adjust*bw*m, where m
varies acccording to the chosen kernel.}

\item{support}{the support of the data. Must be compatible with the supplied
\code{x} and the supplied \code{start} and \code{kernel}}

\item{na.rm}{should \code{NA} values be removed?}

\item{normalized}{should the density be normalized?}

\item{bw}{a bandwidth function (NOT IMPLEMENTED) or a double}

\item{kernel}{the kernel function, chosen from a list of alternatives}

\item{start}{choice of parametric start. Can either be chosen from a vector
of strings, or be supplied via a list containing a \code{density} function
and an \code{estimator} function.}
}
\value{
A function of class kdensity.
}
\description{
Kernel density estimator with parametric start
}
\details{
fill in
}
