% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{listmerge}
\alias{listmerge}
\title{Merges two lists.}
\usage{
listmerge(x, y, type = c("merge", "template"))
}
\arguments{
\item{x}{A list of default arguments.}

\item{y}{A list of supplied arguments}

\item{type}{If \code{merge}, the list will be merge with y
having priority; if \code{template}, named the elements of
y not in x will be discarded after merging.}
}
\value{
A merged list where conflicts are solved in favour
of y. Does not preserve ordering.
}
\description{
Merges two lists.
}
\keyword{internal}
