% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtin_starts_custom_maximum_likelihood.R
\name{mlgamma}
\alias{mlgamma}
\title{Fast maximum likelihood for the gamma distribution}
\usage{
mlgamma(x, rel_eps = 10^-10, max_iter = 100)
}
\arguments{
\item{x}{The data, contained in c(0, Inf)}

\item{rel_eps}{Relative epsilon comparison criterion.}

\item{max_iter}{Maximal number of Newton-Raphson iteratons.}
}
\value{
A vector of two parameters, \code{shape} and \code{rate}.
}
\description{
Fast maximum likelihood for the gamma distribution
}
\references{
Choi, S. C, and R. Wette. "Maximum likelihood estimation of the parameters of the gamma distribution and their bias." Technometrics 11.4 (1969): 683-690.
}
