% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_helper}
\alias{plot_helper}
\title{Helper function for the plot methods}
\usage{
plot_helper(x, range = NULL, plot_start = FALSE, zero_line = TRUE,
  type = c("plot", "lines", "points"), ...)
}
\arguments{
\item{x}{a \code{kdensity} object.}

\item{range}{range of x values.}

\item{plot_start}{logical; if \code{TRUE}, plots the parametric start instead of the kernel density estimate.}

\item{type}{the kind of plot to make}

\item{...}{further plotting parameters.}
}
\value{
None.
}
\description{
Helper function for the plot methods
}
