% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_helper}
\alias{plot_helper}
\title{Helper function for the plot methods.}
\usage{
plot_helper(
  x,
  range = NULL,
  plot_start = FALSE,
  zero_line = TRUE,
  ptype = c("plot", "lines", "points"),
  ...
)
}
\arguments{
\item{x}{A \code{kdensity} object.}

\item{range}{An optional range vector; like \code{x} in \code{plot.default}.}

\item{plot_start}{Logical; if \code{TRUE}, plots the parametric start only.}

\item{zero_line}{Logical; if \code{TRUE}, adds a line at \code{y = 0}.}

\item{ptype}{The kind of plot to make}

\item{...}{Passed to plot.default.}
}
\value{
None.
}
\description{
A helper function for the plot methods that does most of the work under
the hood.
}
\keyword{internal}
