% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_kernel_start_support.R
\name{support_compatible}
\alias{support_compatible}
\title{Checks compatibility between supports.}
\usage{
support_compatible(kernel, start, support)
}
\description{
The supplied support must never be larger than the support of
the parametric start / kernel.
}
